Fix linefeeds after Content-Type header and content.
There has to be an extra linefeed (empty line) between
the multipart content headers and the content itself.
Also, there has to be a new line before the ending
boundary.

The directory structure for entries <year>/<month>/<day>
was removed from README, as it is not meaningfull in
the current development of the engine. It is possible
future format though, for simplified "by hand"
file browsing.